Don't show stop button on job output view if no @deploy there. Otherwise it'll delete the project! :)

+++ app/helpers/deploys_helper.rb
@@ -87,6 +87,7 @@ module DeploysHelper
   end
 
   def stop_button(options = {})
+    return unless @project && @deploy
     link_to "Stop", [@project, @deploy], options.merge({ method: :delete, class: options.fetch(:class, 'btn btn-danger btn-xl') })
   end
